Change: Error handling shared for backend and frontend
