--- backend/src/models/User.js
@@ -1,163 +1,0 @@
-"use strict";
-var __assign = (this && this.__assign) || function () {
-    __assign = Object.assign || function(t) {
-        for (var s, i = 1, n = arguments.length; i < n; i++) {
-            s = arguments[i];
-            for (var p in s) if (Object.prototype.hasOwnProperty.call(s, p))
-                t[p] = s[p];
-        }
-        return t;
-    };
-    return __assign.apply(this, arguments);
-};
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
-var __generator = (this && this.__generator) || function (thisArg, body) {
-    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g = Object.create((typeof Iterator === "function" ? Iterator : Object).prototype);
-    return g.next = verb(0), g["throw"] = verb(1), g["return"] = verb(2), typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
-    function verb(n) { return function (v) { return step([n, v]); }; }
-    function step(op) {
-        if (f) throw new TypeError("Generator is already executing.");
-        while (g && (g = 0, op[0] && (_ = 0)), _) try {
-            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [op[0] & 2, t.value];
-            switch (op[0]) {
-                case 0: case 1: t = op; break;
-                case 4: _.label++; return { value: op[1], done: false };
-                case 5: _.label++; y = op[1]; op = [0]; continue;
-                case 7: op = _.ops.pop(); _.trys.pop(); continue;
-                default:
-                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
-                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
-                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
-                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
-                    if (t[2]) _.ops.pop();
-                    _.trys.pop(); continue;
-            }
-            op = body.call(thisArg, _);
-        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
-        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
-    }
-};
-var __rest = (this && this.__rest) || function (s, e) {
-    var t = {};
-    for (var p in s) if (Object.prototype.hasOwnProperty.call(s, p) && e.indexOf(p) < 0)
-        t[p] = s[p];
-    if (s != null && typeof Object.getOwnPropertySymbols === "function")
-        for (var i = 0, p = Object.getOwnPropertySymbols(s); i < p.length; i++) {
-            if (e.indexOf(p[i]) < 0 && Object.prototype.propertyIsEnumerable.call(s, p[i]))
-                t[p[i]] = s[p[i]];
-        }
-    return t;
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-// backend/src/models/User.ts (UPDATED - Reflecting latest user.interface.ts changes)
-var mongoose_1 = require("mongoose");
-var bcrypt = require("bcrypt");
-;
-var UserSchema = new mongoose_1.Schema({
-    username: {
-        type: String,
-        required: [true, 'Username is required'],
-        unique: true,
-        trim: true,
-    },
-    email: {
-        type: String,
-        required: [true, 'Email is required'],
-        unique: true,
-        lowercase: true,
-        trim: true,
-    },
-    passwordHash: {
-        type: String,
-        required: [true, 'Password is required'],
-    },
-    firstName: {
-        type: String,
-        required: [true, 'First name is required'],
-        trim: true,
-    },
-    familyName: {
-        type: String,
-        required: [true, 'Family name is required'],
-        trim: true,
-    },
-    profilePictureUrl: {
-        type: String,
-        // Removed `default: 'default_profile.png'` here if you're not enforcing a default in the DB directly
-        // The `profilePictureUrl?: string;` in the interface already implies it can be undefined or null.
-        // If you want a default if the user doesn't provide one, keep the default in the schema.
-        // Let's keep the default for now, as it's common to have a fallback image.
-        default: 'default_profile.png',
-        // No 'required: true' here, reflecting it being optional in the interface
-    },
-    bio: {
-        type: String,
-        required: [true, 'Bio is required'], // CHANGED: Now mandatory
-        maxLength: [250, 'Bio cannot exceed 250 characters'],
-    },
-    registrationDate: {
-        type: Date,
-        default: Date.now,
-    },
-    lastLoginDate: {
-        type: Date,
-    },
-    isActive: {
-        type: Boolean,
-        default: true,
-    },
-}, {
-    timestamps: true,
-    toJSON: {
-        transform: function (doc, ret) {
-            // 1. Destructure and omit sensitive fields
-            var passwordHash = ret.passwordHash, _id = ret._id, userObject = __rest(ret, ["passwordHash", "_id"]);
-            // 2. Reconstruct the object for the public API
-            var publicUser = __assign(__assign({}, userObject), { _id: _id.toString() });
-            return publicUser;
-        }
-    }
-});
-UserSchema.pre('save', function (next) {
-    return __awaiter(this, void 0, void 0, function () {
-        var salt, _a;
-        return __generator(this, function (_b) {
-            switch (_b.label) {
-                case 0:
-                    if (!this.isModified('passwordHash')) {
-                        return [2 /*return*/, next()];
-                    }
-                    return [4 /*yield*/, bcrypt.genSalt(10)];
-                case 1:
-                    salt = _b.sent();
-                    _a = this;
-                    return [4 /*yield*/, bcrypt.hash(this.passwordHash, salt)];
-                case 2:
-                    _a.passwordHash = _b.sent();
-                    next();
-                    return [2 /*return*/];
-            }
-        });
-    });
-});
-UserSchema.methods.matchPassword = function (enteredPassword) {
-    return __awaiter(this, void 0, void 0, function () {
-        return __generator(this, function (_a) {
-            switch (_a.label) {
-                case 0: return [4 /*yield*/, bcrypt.compare(enteredPassword, this.passwordHash)];
-                case 1: return [2 /*return*/, _a.sent()];
-            }
-        });
-    });
-};
-var User = mongoose_1.default.model('User', UserSchema);
-exports.default = User;
